feat: refactored Grid and Cell names and removed some unused code, fixing a bug with the Image color
