+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/boat_rental/AddBoatRentalForm.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.key
 import androidx.compose.ui.semantics.Role
 import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
-import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
+import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconCell
 
 class AddBoatRentalForm : AImageListQuestForm<BoatRental, List<BoatRental>>() {
 
@@ -18,7 +18,7 @@ class AddBoatRentalForm : AImageListQuestForm<BoatRental, List<BoatRental>>() {
 
     override val itemContent = @Composable { item: ImageListItem<BoatRental>, index: Int, onClick: () -> Unit, role: Role ->
         key(item.item) {
-            SelectableIconItem(item = item.item,
+            SelectableIconCell(item = item.item,
                 isSelected = item.checked,
                 onClick = onClick,
                 role = role)
