+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/incline_direction/AddInclineForm.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.semantics.Role
 import de.westnordost.streetcomplete.data.osm.geometry.ElementPolylinesGeometry
 import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
-import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
+import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconCell
 import de.westnordost.streetcomplete.util.math.getOrientationAtCenterLineInDegrees
 
 class AddInclineForm : AImageListQuestForm<Incline, Incline>() {
@@ -22,7 +22,7 @@ class AddInclineForm : AImageListQuestForm<Incline, Incline>() {
     override val itemContent =
         @Composable { item: ImageListItem<Incline>, index: Int, onClick: () -> Unit, role: Role ->
             key(item.item to (wayRotation - mapRotation)) {
-                SelectableIconItem(
+                SelectableIconCell(
                     item = item.item,
                     isSelected = item.checked,
                     onClick = onClick,
