+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/oneway/AddOnewayForm.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.semantics.Role
 import de.westnordost.streetcomplete.data.osm.geometry.ElementPolylinesGeometry
 import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
-import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
+import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconCell
 import de.westnordost.streetcomplete.util.math.getOrientationAtCenterLineInDegrees
 
 class AddOnewayForm : AImageListQuestForm<OnewayAnswer, OnewayAnswer>() {
@@ -23,7 +23,7 @@ class AddOnewayForm : AImageListQuestForm<OnewayAnswer, OnewayAnswer>() {
     override val itemContent =
         @Composable { item: ImageListItem<OnewayAnswer>, index: Int, onClick: () -> Unit, role: Role ->
             key(item.item to (wayRotation - mapRotation)) {
-                SelectableIconItem(
+                SelectableIconCell(
                     item = item.item,
                     isSelected = item.checked,
                     onClick = onClick,
