+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/roof_shape/AddRoofShapeForm.kt
@@ -9,14 +9,14 @@ import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 import de.westnordost.streetcomplete.quests.roof_shape.RoofShape.MANY
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
-import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
+import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconCell
 
 class AddRoofShapeForm : AImageListQuestForm<RoofShape, RoofShape>() {
 
     override val itemContent =
         @Composable { item: ImageListItem<RoofShape>, index: Int, onClick: () -> Unit, role: Role ->
             key(item.item) {
-                SelectableIconItem(
+                SelectableIconCell(
                     item = item.item,
                     isSelected = item.checked,
                     onClick = onClick,
