+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/sport/AddSportForm.kt
@@ -10,7 +10,7 @@ import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.quests.AnswerItem
 import de.westnordost.streetcomplete.quests.sport.Sport.MULTI
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
-import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
+import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconCell
 
 class AddSportForm : AImageListQuestForm<Sport, List<Sport>>() {
 
@@ -25,7 +25,7 @@ class AddSportForm : AImageListQuestForm<Sport, List<Sport>>() {
     override val itemContent =
         @Composable { item: ImageListItem<Sport>, index: Int, onClick: () -> Unit, role: Role ->
             key(item.item) {
-                SelectableIconItem(
+                SelectableIconCell(
                     item = item.item,
                     isSelected = item.checked,
                     onClick = onClick,
