+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/tourism_information/AddInformationForm.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.key
 import androidx.compose.ui.semantics.Role
 import de.westnordost.streetcomplete.quests.AImageListQuestForm
 import de.westnordost.streetcomplete.ui.common.image_select.ImageListItem
-import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconItem
+import de.westnordost.streetcomplete.ui.common.image_select.SelectableIconCell
 
 class AddInformationForm : AImageListQuestForm<TourismInformation, TourismInformation>() {
 
@@ -17,7 +17,7 @@ class AddInformationForm : AImageListQuestForm<TourismInformation, TourismInform
     override val itemContent =
         @Composable { item: ImageListItem<TourismInformation>, index: Int, onClick: () -> Unit, role: Role ->
             key(item.item) {
-                SelectableIconItem(
+                SelectableIconCell(
                     item = item.item,
                     isSelected = item.checked,
                     onClick = onClick,
