+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/ui/common/image_select/SelectableIconCell.kt
@@ -6,13 +6,10 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.selection.selectable
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.ContentAlpha
-import androidx.compose.material.LocalContentColor
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -23,14 +20,16 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.asImageBitmap
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
-import de.westnordost.streetcomplete.osm.building.BuildingType
-import de.westnordost.streetcomplete.osm.building.asItem
+import de.westnordost.streetcomplete.quests.boat_rental.BoatRental
+import de.westnordost.streetcomplete.quests.boat_rental.asItem
 import de.westnordost.streetcomplete.ui.ktx.conditional
 import de.westnordost.streetcomplete.ui.theme.SelectionColor
 import de.westnordost.streetcomplete.ui.theme.SelectionFrameColor
@@ -41,7 +40,7 @@ import de.westnordost.streetcomplete.view.ResText
 import de.westnordost.streetcomplete.view.image_select.DisplayItem
 
 @Composable
-fun <T> SelectableIconRightItem(
+fun <T> SelectableIconCell(
     item: DisplayItem<T>,
     isSelected: Boolean,
     onClick: () -> Unit,
@@ -71,13 +70,6 @@ fun <T> SelectableIconRightItem(
             null -> ""
         }
     }
-    val description = remember(item.description) {
-        when (item.description) {
-            is ResText -> context.getString((item.description as ResText).resId)
-            is CharSequenceText -> (item.description as CharSequenceText).text.toString()
-            null -> null
-        }
-    }
 
     Box(
         modifier = modifier
@@ -87,53 +79,46 @@ fun <T> SelectableIconRightItem(
                 role = role
             )
             .conditional(isSelected) {
-                border(4.dp, SelectionFrameColor, RoundedCornerShape(16.dp))
+                border(4.dp, SelectionFrameColor, RoundedCornerShape(8.dp))
             }
+            .padding(4.dp)
             .background(
                 color = SelectionColor.copy(alpha = animatedAlpha),
                 shape = RoundedCornerShape(16.dp)
-            )
-        ,
-        contentAlignment = Alignment.CenterStart
+            ),
+        contentAlignment = Alignment.Center
     ) {
-        Row(
-            modifier = Modifier.padding(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
             imageBitmap?.let { bitmap ->
                 Image(
                     bitmap = bitmap,
-                    contentDescription = title
+                    contentDescription = title,
+                    modifier = Modifier
+                        .wrapContentSize(Alignment.Center),
+                    contentScale = ContentScale.Inside,
+                    alignment = Alignment.Center
                 )
             }
-            Column(
-                horizontalAlignment = Alignment.Start,
-                modifier = Modifier.padding(start = 8.dp)
-            ) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.body1
-                )
-                if (description != null) {
-                    Text(
-                        text = description,
-                        modifier = Modifier.padding(top = 4.dp),
-                        style = MaterialTheme.typography.body2,
-                        color = LocalContentColor.current.copy(alpha = ContentAlpha.medium),
-                    )
-                }
-            }
+
+            Text(
+                text = title,
+                textAlign = TextAlign.Center,
+                style = MaterialTheme.typography.body1,
+                modifier = Modifier
+                    .align(Alignment.CenterHorizontally)
+                    .padding(8.dp)
+            )
         }
     }
 }
 
 @Composable
 @Preview(showBackground = true)
-fun SelectableIconRightItemPreview() {
+fun SelectableIconCellPreview() {
     var selected by remember { mutableStateOf(false) }
 
-    SelectableIconRightItem(
-        item = BuildingType.APARTMENTS.asItem(),
+    SelectableIconCell(
+        item = BoatRental.CANOE.asItem(),
         isSelected = selected,
         onClick = { selected = !selected }
     )
