+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/ui/common/image_select/SelectableIconRightCell.kt
@@ -5,10 +5,14 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.selection.selectable
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.ContentAlpha
+import androidx.compose.material.LocalContentColor
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -17,25 +21,16 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shadow
 import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
-import de.westnordost.streetcomplete.quests.bicycle_repair_station.BicycleRepairStationService
-import de.westnordost.streetcomplete.quests.bicycle_repair_station.asItem
+import de.westnordost.streetcomplete.osm.building.BuildingType
+import de.westnordost.streetcomplete.osm.building.asItem
 import de.westnordost.streetcomplete.ui.ktx.conditional
-import de.westnordost.streetcomplete.ui.ktx.pxToDp
 import de.westnordost.streetcomplete.ui.theme.SelectionColor
 import de.westnordost.streetcomplete.ui.theme.SelectionFrameColor
 import de.westnordost.streetcomplete.view.CharSequenceText
@@ -45,7 +40,7 @@ import de.westnordost.streetcomplete.view.ResText
 import de.westnordost.streetcomplete.view.image_select.DisplayItem
 
 @Composable
-fun <T> SelectableImageItem(
+fun <T> SelectableIconRightCell(
     item: DisplayItem<T>,
     isSelected: Boolean,
     onClick: () -> Unit,
@@ -56,7 +51,6 @@ fun <T> SelectableImageItem(
         targetValue = if (isSelected) 0.5f else 0f,
         label = "OverlayAlpha"
     )
-
     val context = LocalContext.current
     val imageBitmap = remember(item.image) {
         when (item.image) {
@@ -69,7 +63,6 @@ fun <T> SelectableImageItem(
             else -> null
         }
     }
-
     val title = remember(item.title) {
         when (item.title) {
             is ResText -> context.getString((item.title as ResText).resId)
@@ -77,8 +70,13 @@ fun <T> SelectableImageItem(
             null -> ""
         }
     }
-
-    var imageWidth by remember { mutableStateOf(0) }
+    val description = remember(item.description) {
+        when (item.description) {
+            is ResText -> context.getString((item.description as ResText).resId)
+            is CharSequenceText -> (item.description as CharSequenceText).text.toString()
+            null -> null
+        }
+    }
 
     Box(
         modifier = modifier
@@ -87,44 +85,39 @@ fun <T> SelectableImageItem(
                 onClick = onClick,
                 role = role
             )
+            .conditional(isSelected) {
+                border(4.dp, SelectionFrameColor, RoundedCornerShape(16.dp))
+            }
+            .background(
+                color = SelectionColor.copy(alpha = animatedAlpha),
+                shape = RoundedCornerShape(16.dp)
+            ),
+        contentAlignment = Alignment.CenterStart
     ) {
-        Box(
-            modifier = Modifier
-                .conditional(isSelected) {
-                    border(4.dp, SelectionFrameColor)
-                }
-                .padding(4.dp)
-                .background(SelectionColor.copy(alpha = animatedAlpha))
+        Row(
+            modifier = Modifier.padding(8.dp),
+            verticalAlignment = Alignment.CenterVertically
         ) {
             imageBitmap?.let { bitmap ->
                 Image(
                     bitmap = bitmap,
-                    contentDescription = title,
-                    modifier = Modifier
-                        .align(Alignment.Center)
-                        .onSizeChanged {
-                            imageWidth = it.width
-                        }
+                    contentDescription = title
                 )
-
-                if (imageWidth > 0) {
+            }
+            Column(
+                horizontalAlignment = Alignment.Start,
+                modifier = Modifier.padding(start = 8.dp)
+            ) {
+                Text(
+                    text = title,
+                    style = MaterialTheme.typography.body1
+                )
+                if (description != null) {
                     Text(
-                        text = title,
-                        style = TextStyle(
-                            color = Color.White,
-                            fontSize = 12.sp,
-                            fontWeight = FontWeight.Bold,
-                            shadow = Shadow(
-                                color = Color.Black,
-                                offset = Offset(2f, 2f),
-                                blurRadius = 4f
-                            )
-                        ),
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier
-                            .align(Alignment.BottomCenter)
-                            .width(imageWidth.pxToDp())
-                            .padding(8.dp)
+                        text = description,
+                        modifier = Modifier.padding(top = 4.dp),
+                        style = MaterialTheme.typography.body2,
+                        color = LocalContentColor.current.copy(alpha = ContentAlpha.medium),
                     )
                 }
             }
@@ -132,14 +125,13 @@ fun <T> SelectableImageItem(
     }
 }
 
-
 @Composable
 @Preview(showBackground = true)
-fun SelectableImagePreview() {
+fun SelectableIconRightCellPreview() {
     var selected by remember { mutableStateOf(false) }
 
-    SelectableImageItem(
-        item = BicycleRepairStationService.CHAIN_TOOL.asItem(),
+    SelectableIconRightCell(
+        item = BuildingType.APARTMENTS.asItem(),
         isSelected = selected,
         onClick = { selected = !selected }
     )
