+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/ui/common/image_select/SelectableImageCell.kt
@@ -5,32 +5,39 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.selection.selectable
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.graphics.Shadow
 import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
-import de.westnordost.streetcomplete.quests.boat_rental.BoatRental
-import de.westnordost.streetcomplete.quests.boat_rental.asItem
+import de.westnordost.streetcomplete.quests.bicycle_repair_station.BicycleRepairStationService
+import de.westnordost.streetcomplete.quests.bicycle_repair_station.asItem
 import de.westnordost.streetcomplete.ui.ktx.conditional
+import de.westnordost.streetcomplete.ui.ktx.pxToDp
 import de.westnordost.streetcomplete.ui.theme.SelectionColor
 import de.westnordost.streetcomplete.ui.theme.SelectionFrameColor
 import de.westnordost.streetcomplete.view.CharSequenceText
@@ -40,7 +47,7 @@ import de.westnordost.streetcomplete.view.ResText
 import de.westnordost.streetcomplete.view.image_select.DisplayItem
 
 @Composable
-fun <T> SelectableIconItem(
+fun <T> SelectableImageCell(
     item: DisplayItem<T>,
     isSelected: Boolean,
     onClick: () -> Unit,
@@ -51,6 +58,7 @@ fun <T> SelectableIconItem(
         targetValue = if (isSelected) 0.5f else 0f,
         label = "OverlayAlpha"
     )
+
     val context = LocalContext.current
     val imageBitmap = remember(item.image) {
         when (item.image) {
@@ -63,6 +71,7 @@ fun <T> SelectableIconItem(
             else -> null
         }
     }
+
     val title = remember(item.title) {
         when (item.title) {
             is ResText -> context.getString((item.title as ResText).resId)
@@ -78,47 +87,52 @@ fun <T> SelectableIconItem(
                 onClick = onClick,
                 role = role
             )
-            .conditional(isSelected) {
-                border(4.dp, SelectionFrameColor, RoundedCornerShape(8.dp))
-            }
-            .padding(4.dp)
-            .background(
-                color = SelectionColor.copy(alpha = animatedAlpha),
-                shape = RoundedCornerShape(16.dp)
-            ),
-        contentAlignment = Alignment.Center
     ) {
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+        Box(
+            modifier = Modifier
+                .conditional(isSelected) {
+                    border(4.dp, SelectionFrameColor)
+                }
+                .padding(4.dp)
+        ) {
             imageBitmap?.let { bitmap ->
                 Image(
                     bitmap = bitmap,
                     contentDescription = title,
                     modifier = Modifier
-                        .wrapContentSize(Alignment.Center),
-                    contentScale = ContentScale.Inside,
-                    alignment = Alignment.Center
+                        .align(Alignment.Center),
+                    colorFilter = ColorFilter.tint(SelectionColor.copy(alpha = animatedAlpha),
+                        BlendMode.Color)
+                )
+                Text(
+                    text = title,
+                    style = TextStyle(
+                        color = Color.White,
+                        fontSize = 12.sp,
+                        fontWeight = FontWeight.Bold,
+                        shadow = Shadow(
+                            color = Color.Black,
+                            offset = Offset(2f, 2f),
+                            blurRadius = 4f
+                        )
+                    ),
+                    textAlign = TextAlign.Center,
+                    modifier = Modifier
+                        .align(Alignment.BottomCenter)
+                        .padding(8.dp)
                 )
             }
-
-            Text(
-                text = title,
-                textAlign = TextAlign.Center,
-                style = MaterialTheme.typography.body1,
-                modifier = Modifier
-                    .align(Alignment.CenterHorizontally)
-                    .padding(8.dp)
-            )
         }
     }
 }
 
 @Composable
 @Preview(showBackground = true)
-fun SelectableIconPreview() {
+fun SelectableImageCellPreview() {
     var selected by remember { mutableStateOf(false) }
 
-    SelectableIconItem(
-        item = BoatRental.CANOE.asItem(),
+    SelectableImageCell(
+        item = BicycleRepairStationService.CHAIN_TOOL.asItem(),
         isSelected = selected,
         onClick = { selected = !selected }
     )
